lambda/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
gspread==6.1.4
google-auth==2.37.0
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
